update setup script

+++ setup.sh
@@ -2,14 +2,21 @@
 
 mkdir -p ~/.config/nvim
 
-CURR_DIR = $PWD
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 
-ln -s $CURR_DIR/nvim/init.vim ~/.config/nvim/init.vim
+# link nvim config
+ln -sf $SCRIPT_DIR/nvim/init.vim ~/.config/nvim/init.vim
 
-ln -s $CURR_DIR/gitconfig ~/.gitconfig
-cp gitconfig-work ~/.gitconfig-work
-ln -s $CURR_DIR/gitconfig-projects ~/.gitconfig-projects
-ln -s $CURR_DIR/gitexcludes ~/.gitexcludes
+# link git configs
+ln -sf $SCRIPT_DIR/gitexcludes ~/.gitexcludes
+ln -sf $SCRIPT_DIR/gitconfig-projects ~/.gitconfig-projects
+if [ -f ~/.gitconfig-work ];
+then
+  echo ".gitconfig-work already present, omitting update.";
+else
+  cp gitconfig-work ~/.gitconfig-work;
+fi
+ln -sf $SCRIPT_DIR/gitconfig ~/.gitconfig
 
-ln -s $CURR_DIR/zshrc.custom ~/.zshrc.custom
-
+# link zsh config
+ln -sf $SCRIPT_DIR/zshrc.custom ~/.zshrc.custom
